Store timedelta coords as float64 to fix NaT decode overflow
Forecast lead_time and the derived forecast-length coords were encoded as
int64 seconds. xarray writes NaT as the int64 sentinel (np.iinfo(int64).min),
and decoding that sentinel to a finer resolution (seconds -> us/ns) multiplies
it and overflows int64 on some xarray versions (OutOfBoundsTimedelta /
"Overflow in int64 * timedelta64 multiplication"). The int64[us] decode hint
added in the pandas 3 update made this surface for downstream readers.

No int64 fill sentinel can avoid this: the mask coder maps the fill to
int64.min before the timedelta range check runs. Storing the coords as float64
makes NaT a plain NaN that decodes correctly across xarray versions. CF places
no integer requirement on time/forecast_period encoding, and the in-memory
dtype for readers is unchanged (timedelta64[us], driven by the dtype attribute).
Regenerated every forecast template.

Tests: load every checked-in template's coords (regression for the overflow,
exercising stored NaT sentinels) and assert every timedelta coord is stored as
float64 (cross-template consistency guard). Allow all-fill coords (e.g. all-NaT
ingested_forecast_length) to write no chunk.

+++ tests/noaa/gefs/forecast_35_day/template_config_test.py
@@ -245,7 +245,7 @@ def test_coordinates_configuration(
 
     # Test lead_time coordinate
     lead_time_coord = next(coord for coord in coords if coord.name == "lead_time")
-    assert lead_time_coord.encoding.dtype == "int64"
+    assert lead_time_coord.encoding.dtype == "float64"
     assert lead_time_coord.encoding.units == "seconds"
     # Should chunk all lead times together
     assert lead_time_coord.encoding.chunks == 181
